Fikser stylingfeil og endrer mer korrekt bruk av ds-komponenter.
