+++ package.json
@@ -46,7 +46,6 @@
     "nav-frontend-ekspanderbartpanel-style": "^2.0.2",
     "nav-frontend-etiketter-style": "^2.0.3",
     "nav-frontend-hjelpetekst-style": "^3.0.2",
-    "nav-frontend-js-utils": "^1.0.20",
     "nav-frontend-knapper-style": "^2.1.2",
     "nav-frontend-lenkepanel-style": "^2.0.2",
     "nav-frontend-lenker-style": "^2.0.2",
@@ -54,7 +53,6 @@
     "nav-frontend-lukknapp-style": "^2.0.2",
     "nav-frontend-modal-style": "^2.0.2",
     "nav-frontend-paneler-style": "^2.0.2",
-    "nav-frontend-popover": "2.0.4",
     "nav-frontend-popover-style": "2.0.2",
     "nav-frontend-skjema-style": "^3.0.3",
     "nav-frontend-spinner-style": "^1.0.2",
