+++ src/components/infotekst/Infotekst.less
@@ -12,7 +12,7 @@
     &:focus {
       .infotekst__title-i-icon {
         outline: 0;
-        box-shadow: 0 0 0 3px @fokusFarge;
+        box-shadow: 0 0 0 3px var(--navds-global-color-focus);
       }
     }
   }
