+++ src/components/kilde/Kilde.less
@@ -22,7 +22,10 @@
   }
   &__tekst {
     padding: 0.5rem 0;
-    color: @navGra60;
+    color: var(--navds-color-gray-900);
     text-transform: uppercase;
+  }
+  &__knapp.navds-button {
+    margin-left: -1.5rem;
   }
 }
