+++ src/index.less
@@ -5,27 +5,9 @@
 @maxMobile: ~"all and (max-width: 420px)";
 @minMobile: ~"all and (min-width: 420px)";
 
-@import (reference) "~nav-frontend-core/less/_variabler";
+@import (reference) "~@navikt/ds-css/variables.css";
 @import (less) "~react-day-picker/lib/style.css";
-@import "~nav-frontend-spinner-style";
-@import "~nav-frontend-typografi-style";
-@import "~nav-frontend-ekspanderbartpanel-style";
-@import "~nav-frontend-paneler-style";
-@import "~nav-frontend-lesmerpanel-style";
-@import "~nav-frontend-chevron-style";
-@import "~nav-frontend-lenker-style";
-@import "~nav-frontend-lukknapp-style";
-@import "~nav-frontend-skjema-style";
-@import "~nav-frontend-modal-style";
-@import "~nav-frontend-veileder-style";
-@import "~nav-frontend-tabs-style";
-@import "~nav-frontend-hjelpetekst-style";
-@import "~nav-frontend-etiketter-style";
-@import "~nav-frontend-veilederpanel-style";
-@import "~nav-frontend-alertstriper-style";
-@import "~nav-frontend-knapper-style";
-@import "~nav-frontend-lenkepanel-style";
-@import "~nav-frontend-popover-style";
+@import "~@navikt/ds-css/index.css";
 
 @import "App.less";
 @import "pages/forside/sections/3-header/Header.less";
