+++ src/pages/endre-personopplysninger/EndreOpplysninger.less
@@ -1,6 +1,6 @@
 .endreOpplysninger {
   &__page {
-    background-color: @navGraBakgrunn;
+    background-color: var(--navds-global-color-background);
     position: absolute;
     top: 0;
     left: 0;
