+++ src/pages/forside/sections/4-personinfo/2-kontaktinfo/subsections/telefonnummer/Telefonnummer.less
@@ -7,11 +7,17 @@
     border-bottom: 1px solid #c6c2bf;
   }
   &__leggtil {
-    padding-left: 0;
-    margin: 0.5rem 0;
-    align-items: center;
-    display: flex;
-    border: 0;
+    &.navds-button {
+      display: none;
+      margin: 0.5rem 0 0.5rem -1.3rem;
+      align-items: center;
+      display: flex;
+      border: 0;
+    }
+
+    .kilde__icon {
+      display: flex;
+    }
   }
   &__divider {
     height: 1px;
@@ -50,22 +56,6 @@
   &__knapper {
     display: flex;
     margin-top: 1rem;
-  }
-  &__knapp.navds-button {
-    margin-left: 0.25rem;
-  }
-  &__knapp-med-ikon {
-    justify-content: center;
-    align-items: center;
-    padding: 0.275rem !important;
-
-    .navds-label {
-      display: inherit;
-    }
-
-    &:last-child {
-      margin-left: 0.25rem;
-    }
   }
   &__knapp-ikon,
   &__knapp-tekst {
