+++ src/pages/forside/sections/4-personinfo/3-adresser/Adresser.less
@@ -20,11 +20,18 @@
       }
     }
   }
+  &__slett-kontaktadresse.navds-button {
+    margin-left: -1.5rem;
+  }
   &__tittel {
     padding: 0 0 1rem 0;
   }
   &__lesmer {
     padding: 0;
+
+    .navds-read-more__button {
+      margin-left: -0.5rem;
+    }
   }
   &__linje {
     width: 100%;
