+++ src/pages/forside/sections/4-personinfo/5-ditt-nav-kontor/DittNavKontor.less
@@ -22,6 +22,7 @@
   &__adresser {
     display: flex;
     flex-wrap: wrap;
+    margin-top: 1rem;
   }
   &__publikumsmottak,
   &__postadresse {
