+++ src/pages/forside/sections/7-mer-informasjon/MerInformasjon.less
@@ -10,24 +10,8 @@
     flex-direction: column;
   }
 
-  &__button {
-    display: flex;
-    justify-content: left;
-    text-align: left;
-    background: none;
-    border: 0;
-    padding: 1rem 0;
-    width: 100%;
-    border-bottom: 1px solid @navGra60;
-    &:focus {
-      background: @fokusFarge;
-      .mi__lenke {
-        color: white;
-      }
-    }
-    &:last-child {
-      border-bottom: 0;
-    }
+  &__enkeltpanel {
+    margin-bottom: 1rem;
   }
 
   &__overskrift {
