+++ src/pages/forside/sections/7-mer-informasjon/MerInformasjon.tsx
@@ -1,28 +1,34 @@
 import React, { Fragment } from "react";
-import { FormattedMessage } from "react-intl";
-import MerInformasjonPanel from "./MerInformasjonPanel";
+import { FormattedMessage, useIntl } from "react-intl";
+import MerInformasjonPanelinnhold from "./MerInformasjonPanelinnhold";
 import merInformasjon from "./MerInformasjonData";
-import { Heading, Panel } from "@navikt/ds-react";
+import { Heading, Panel, ReadMore } from "@navikt/ds-react";
 
-const AlternativListe = () => (
-  <Panel className="el__panel">
-    <div className="mi__content">
-      <div className="mi__overskrift">
-        <Heading size={"medium"} level={"2"}>
-          <FormattedMessage id="alternativer.tittel" />
-        </Heading>
+const AlternativListe = () => {
+  const { formatMessage: msg } = useIntl();
+
+  return (
+    <Panel className="el__panel">
+      <div className="mi__content">
+        <div className="mi__overskrift">
+          <Heading size={"medium"} level={"2"}>
+            <FormattedMessage id="alternativer.tittel" />
+          </Heading>
+        </div>
+        {merInformasjon.map((info, i) => (
+          <Fragment key={i}>
+            <ReadMore
+              key={info.id}
+              header={msg({ id: info.tittel })}
+              className="mi__enkeltpanel"
+            >
+              <MerInformasjonPanelinnhold melding={info.melding} />
+            </ReadMore>
+          </Fragment>
+        ))}
       </div>
-      {merInformasjon.map((info, i) => (
-        <Fragment key={i}>
-          <MerInformasjonPanel
-            key={info.id}
-            tittel={info.tittel}
-            melding={info.melding}
-          />
-        </Fragment>
-      ))}
-    </div>
-  </Panel>
-);
+    </Panel>
+  );
+};
 
 export default AlternativListe;
